Add GraphRequestAdapter
